requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
pyserial==3.5
pynmea2==1.19.0
python-socketio==5.10.0
simple-websocket==1.0.0
